pyproject.toml
[project]
name = "lia"
version = "0.1.0"
description = "AI Agent to generate an italian text"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fpdf2>=2.8.5",
    "google-genai==1.12.1",
    "pytest>=8.4.2",
    "python-dotenv==1.1.0",
]
